% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodalTest.R
\name{bimodalTest}
\alias{bimodalTest}
\title{bimodalTest}
\usage{
bimodalTest(se, cvExpr, threshold = NULL)
}
\arguments{
\item{se}{A SummarizedExperiment containing the gene expression data.}

\item{cvExpr}{The output from cvExpr.}

\item{threshold}{Optional. The quantile threshold of genes to be used for clustering analaysis. Default is NULL.}
}
\value{
Returns a list containing the output from the bimodal test.
}
\description{
Performs a bimodal test over a genomic region of interest.
}
\examples{
library(GenomicRanges)
data(datExprs)
chr21 <- GRanges("21:1-46709983")
cvExpr.out <- cvExpr(se = datExprs, region = chr21)
bimodalTest.out <- bimodalTest(se = datExprs, cvExpr = cvExpr.out, threshold = "25\%")
}
\author{
Benjamin Mayne
}
