% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprFinder.R
\name{exprFinder}
\alias{exprFinder}
\title{exprFinder}
\usage{
exprFinder(se, seqInfo, binWidth, threshold, threads)
}
\arguments{
\item{se}{A SummarizedExperiment containing the gene expression data.}

\item{seqInfo}{A SeqInfo object containing the lengths of the chromosomes.}

\item{binWidth}{The length of each bin.}

\item{threshold}{Optional. The quantile threshold of expression variation of genes to be used at each bin. Default is NULL.}

\item{threads}{Total number of threads to be used. Default is 1.}
}
\value{
Returns a data frame containing the genomic locations of each bin with bimodality statistics.
}
\description{
Performs a bimodality test at multiple defined bin sizes across the genome.
}
\author{
Benjamin Mayne
}
