% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaExpr.R
\name{pcaExpr}
\alias{pcaExpr}
\title{pcaExprs}
\usage{
pcaExprs(se, cvExpr, threshold = NULL)
}
\arguments{
\item{se}{A SummarizedExperiment containing the gene expression data and the clustering output from clusterExpr.}

\item{threshold}{Optional. The quantile threshold of genes to be used for clustering analaysis. Default is NULL.}

\item{cvExprA}{The output from cvExpr.}
}
\value{
Returns a PCA plot showing the seperation of samples that were labelled hyperploidy or hypoploidy in clusterExpr.
}
\description{
Determines the coefficient of variation (CV) for gene expression
}
\author{
Benjamin Mayne
}
