% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcvExpr.R
\name{plotcvExpr}
\alias{plotcvExpr}
\title{plotcvExpr}
\usage{
plotcvExpr(cvExpr)
}
\value{
Returns a barplot showing the CV for each gene identifier over the region of interest.
}
\description{
Plots the coefficient of variation (CV) for gene expression over the region of interest
}
\author{
Benjamin Mayne
}
