% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodalTest.R
\name{bimodalTest}
\alias{bimodalTest}
\title{bimodalTest}
\usage{
bimodalTest(se, cvExpr, threshold = NULL)
}
\arguments{
\item{se}{A SummarizedExperiment containing the gene expression data.}

\item{threshold}{Optional. The quantile threshold of genes to be used for clustering analaysis. Default is NULL.}

\item{cvExprA}{The output from cvExpr.}
}
\value{
Returns a list containing the output from the bimodal test.
}
\description{
Performs a bimodal
}
\author{
Benjamin Mayne
}
