% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodalTest.R
\name{bimodalTest}
\alias{bimodalTest}
\title{bimodalTest}
\usage{
bimodalTest(se, cvExpr, threshold = NULL)
}
\arguments{
\item{se}{A SummarizedExperiment containing the normalised gene expression data.}

\item{cvExpr}{The output from cvExpr.}

\item{threshold}{Optional. The quantile threshold of genes to be used for clustering analaysis. Default is NULL.}
}
\value{
Returns a list containing the output from the bimodal test.
}
\description{
Performs a bimodal test and calcualtes Hartigan's statistic and p-value over a genomic region of interest using gene expression data set, the output from cvExpr.
}
\details{
Performs a bimodal test and calculates Hartigan's dip test statistic for unimodality for a given gene expression data set.
A bimodality coefficient value > 5/9 suggests bimodality and the closer the bimodality ratio is to 1, the more evenly distrubted the data set.
The dip statistic and p-value can be used to determine if the region of interest is statistically significant.

The second part of the function returns the Z score means which can be used to visualise the denisty or distribution of the samples.
}
\examples{
library(GenomicRanges)
data(datExprs)
chr21 <- GRanges("21:1-46709983")
cvExpr.out <- cvExpr(se = datExprs, region = chr21)
bimodalTest.out <- bimodalTest(se = datExprs, cvExpr = cvExpr.out, threshold = "25\%")
}
\author{
Benjamin Mayne
}
