% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterExpr.R
\name{clusterExpr}
\alias{clusterExpr}
\title{clusterExpr}
\usage{
clusterExpr(se, cvExpr, threshold = NULL)
}
\arguments{
\item{se}{A SummarizedExperiment containing the gene expression data.}

\item{threshold}{Optional. The quantile threshold of genes to be used for clustering analaysis. Default is NULL.}

\item{cvExprA}{The output from cvExpr.}
}
\value{
Returns a SummarizedExperiment containing the original inputted se, but where an additional column labelled Ploidy has been added into the meta data containing the classification of each sample.
}
\description{
Performs a bimodal test using the genes within the region of interest.
}
\author{
Benjamin Mayne
}
