% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvExpr.R
\name{cvExpr}
\alias{cvExpr}
\title{cvExpr}
\usage{
cvExpr(se, region)
}
\arguments{
\item{se}{A SummarizedExperiment containing the gene expression data.}

\item{region}{A GRanges object containing the genomic location of the region of interest. This can either be an entire length or the subset of a chromosome.}
}
\value{
Returns a list containing the CV of each gene and the the quantile threshold of the data.
}
\description{
Determines the coefficient of variation (CV) for gene expression
}
\author{
Benjamin Mayne
}
